python-backend/requirements.txt
fastapi==0.116.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.107.0
langchain==0.1.0
langchain-openai==0.0.5
pydantic==2.5.0
requests==2.31.0
httpx==0.27.2
